🔧 refactor: remove unused selector data method and clean up imports
api/namespaces/files.py: Removed the on_get_selectors_data method as it was unused.
api/routes/bot/__init__.py: Cleaned up imports by removing unused form classes and adjusted the loadform import.
api/routes/bot/botlaunch_methods.py: Added loadform function to handle form loading and processing, including error handling and logging.
